Melhorando a segurança com CartaoResponse
